dedup local spark and py4j backend functionality
