+++ hail/python/hail/backend/py4j_backend.py
@@ -10,9 +10,14 @@ from hail.expr import construct_expr
 from hail.ir import JavaIR, finalize_randomness
 from hail.ir.renderer import CSERenderer
 from hail.utils.java import FatalError, Env
+from hail.expr.blockmatrix_type import tblockmatrix
+from hail.expr.matrix_type import tmatrix
+from hail.expr.table_type import ttable
+from hail.expr.types import dtype
+from hail.ir import finalize_randomness
+from hail.ir.renderer import CSERenderer
+
 from .backend import Backend, fatal_error_from_java_error_triplet
-from ..expr import Expression
-from ..expr.types import HailType
 
 
 def handle_java_exception(f):
@@ -64,33 +69,6 @@ class Py4JBackend(Backend):
     @abc.abstractmethod
     def utils_package_object(self):
         pass
-
-    @abc.abstractmethod
-    def _parse_value_ir(self, code, ref_map={}, ir_map={}):
-        pass
-
-    @abc.abstractmethod
-    def _to_java_value_ir(self, ir):
-        pass
-
-    def register_ir_function(self,
-                             name: str,
-                             type_parameters: Union[Tuple[HailType, ...], List[HailType]],
-                             value_parameter_names: Union[Tuple[str, ...], List[str]],
-                             value_parameter_types: Union[Tuple[HailType, ...], List[HailType]],
-                             return_type: HailType,
-                             body: Expression):
-        r = CSERenderer(stop_at_jir=True)
-        code = r(finalize_randomness(body._ir))
-        jbody = (self._parse_value_ir(code, ref_map=dict(zip(value_parameter_names, value_parameter_types)), ir_map=r.jirs))
-
-        Env.hail().expr.ir.functions.IRFunctionRegistry.pyRegisterIR(
-            name,
-            [ta._parsable_string() for ta in type_parameters],
-            value_parameter_names,
-            [pt._parsable_string() for pt in value_parameter_types],
-            return_type._parsable_string(),
-            jbody)
 
     def execute(self, ir, timed=False):
         jir = self._to_java_value_ir(ir)
@@ -152,6 +130,65 @@ class Py4JBackend(Backend):
 
     def remove_liftover(self, name, dest_reference_genome):
         self._jbackend.pyRemoveLiftover(name, dest_reference_genome)
+
+    def parse_vcf_metadata(self, path):
+        return json.loads(self._jhc.pyParseVCFMetadataJSON(self._jbackend.fs(), path))
+
+    def index_bgen(self, files, index_file_map, referenceGenomeName, contig_recoding, skip_invalid_loci):
+        self._jbackend.pyIndexBgen(files, index_file_map, referenceGenomeName, contig_recoding, skip_invalid_loci)
+
+    def import_fam(self, path: str, quant_pheno: bool, delimiter: str, missing: str):
+        return json.loads(self._jbackend.pyImportFam(path, quant_pheno, delimiter, missing))
+
+    def _to_java_ir(self, ir, parse):
+        if not hasattr(ir, '_jir'):
+            r = CSERenderer(stop_at_jir=True)
+            # FIXME parse should be static
+            ir._jir = parse(r(finalize_randomness(ir)), ir_map=r.jirs)
+        return ir._jir
+
+    def _parse_value_ir(self, code, ref_map={}, ir_map={}):
+        return self._jbackend.parse_value_ir(
+            code,
+            {k: t._parsable_string() for k, t in ref_map.items()},
+            ir_map)
+
+    def _parse_table_ir(self, code, ir_map={}):
+        return self._jbackend.parse_table_ir(code, ir_map)
+
+    def _parse_matrix_ir(self, code, ir_map={}):
+        return self._jbackend.parse_matrix_ir(code, ir_map)
+
+    def _parse_blockmatrix_ir(self, code, ir_map={}):
+        return self._jbackend.parse_blockmatrix_ir(code, ir_map)
+
+    def _to_java_value_ir(self, ir):
+        return self._to_java_ir(ir, self._parse_value_ir)
+
+    def _to_java_table_ir(self, ir):
+        return self._to_java_ir(ir, self._parse_table_ir)
+
+    def _to_java_matrix_ir(self, ir):
+        return self._to_java_ir(ir, self._parse_matrix_ir)
+
+    def _to_java_blockmatrix_ir(self, ir):
+        return self._to_java_ir(ir, self._parse_blockmatrix_ir)
+
+    def value_type(self, ir):
+        jir = self._to_java_value_ir(ir)
+        return dtype(jir.typ().toString())
+
+    def table_type(self, tir):
+        jir = self._to_java_table_ir(tir)
+        return ttable._from_java(jir.typ())
+
+    def matrix_type(self, mir):
+        jir = self._to_java_matrix_ir(mir)
+        return tmatrix._from_java(jir.typ())
+
+    def blockmatrix_type(self, bmir):
+        jir = self._to_java_blockmatrix_ir(bmir)
+        return tblockmatrix._from_java(jir.typ())
 
     @property
     def requires_lowering(self):
